+++ backend/LICENSE
@@ -1,4 +1,4 @@
-Chopper-Backend Licensing
+Choppa-Backend Licensing
 Copyright 2020 Uways Eid, Brett Taylor, Adam Bengey, and Marin Donchev.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -22,7 +22,7 @@ Any license notice or attribution required by the License must also include this
 Commons Clause License Condition notice.
 
 For purposes of the clause above, the “Licensor” are Uways Eid, Brett Taylor, Adam Bengey, and Marin Donchev. 
-The “License” is the Apache License, Version 2.0, and the Software is Chopper-backend (software provided with this
+The “License” is the Apache License, Version 2.0, and the Software is Choppa-backend (software provided with this
 notice).
 
 The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
