+++ backend/config/choppa.config.ts
@@ -1,0 +1,1 @@
+export const port = parseInt(Deno.env.get("PORT") || "8888");
