+++ backend/config/database.config.ts
@@ -11,12 +11,12 @@ interface ConnectionOptions {
 }
 
 /**
- * Chopper's database connection options
+ * Choppa's database connection options
  */
 export const connectionOptions: ConnectionOptions = {
-    user: Deno.env.get("DB_USERNAME") || "chopper",
+    user: Deno.env.get("DB_USERNAME") || "choppa",
     password: Deno.env.get("DB_PASSWORD") || "12345",
-    database: Deno.env.get("DB_NAME") || "chopper",
+    database: Deno.env.get("DB_NAME") || "choppa",
     hostname: Deno.env.get("DB_HOST_URL") || "localhost",
     port: parseInt(Deno.env.get("DB_PORT") || "5432"),
     poolSize: parseInt(Deno.env.get("DB_POOL_CONNECTIONS") || "15")
